requirements.txt
flask
cohere
python-dotenv
apscheduler
requests